Revert "SDK-1300: Move docscanerr to yotierror, to be reused in other services"
This reverts commit f95197b6849b8c31f1efd7a6e4778542c5e7aaf1.
