+++ docscan/response/docscanerr/errors_test.go
@@ -1,4 +1,4 @@
-package yotierror
+package docscanerr
 
 import (
 	"bytes"
