+++ docscan/sandbox/client_test.go
@@ -13,8 +13,8 @@ import (
 	"testing"
 
 	"github.com/getyoti/yoti-go-sdk/v3/cryptoutil"
+	"github.com/getyoti/yoti-go-sdk/v3/docscan/response/docscanerr"
 	"github.com/getyoti/yoti-go-sdk/v3/docscan/sandbox/request"
-	"github.com/getyoti/yoti-go-sdk/v3/yotierror"
 	"gotest.tools/v3/assert"
 )
 
@@ -42,7 +42,7 @@ func TestClient_ConfigureSessionResponse_ShouldReturnErrorIfNotCreated(t *testin
 
 func TestClient_ConfigureSessionResponse_ShouldReturnFormattedErrorWithResponse(t *testing.T) {
 	key, _ := rsa.GenerateKey(rand.Reader, 1024)
-	jsonBytes, _ := json.Marshal(yotierror.DataObject{
+	jsonBytes, _ := json.Marshal(docscanerr.DataObject{
 		Code:    "SOME_CODE",
 		Message: "some message",
 	})
@@ -62,7 +62,7 @@ func TestClient_ConfigureSessionResponse_ShouldReturnFormattedErrorWithResponse(
 	err := client.ConfigureSessionResponse("some_session_id", &request.ResponseConfig{})
 	assert.ErrorContains(t, err, "400: SOME_CODE - some message")
 
-	errorResponse := err.(*yotierror.Error).Response
+	errorResponse := err.(*docscanerr.Error).Response
 	assert.Equal(t, response, errorResponse)
 
 	body, _ := ioutil.ReadAll(errorResponse.Body)
